move DashboardLayout to the Layout folder
